fix:idTokenでユーザー情報を取得する際、取り出す情報を指定できるように＋ファイル分割
